grouped integration tests into one class
